faafo/requirements.txt
#pbr>=1.6
pbr
pytz
positional
iso8601
#anyjson>=0.3.3
#anyjson
#eventlet>=0.17.4
eventlet
#PyMySQL>=0.6.2,<0.7  # 0.7 design change breaks faafo, MIT License
#PyMySQL>=0.6.2
PyMySQL
#SQLAlchemy>1.3,<1.4  # 1.4 breaks faafo list
#SQLAlchemy>=2.0.16
SQLAlchemy
#Pillow==2.4.0 # MIT
Pillow
#requests>=2.5.2
requests
Flask-Bootstrap
Flask
flask-restless-ng
flask-sqlalchemy
#oslo.config>=2.3.0  # Apache-2.0
#oslo.log>=1.8.0  # Apache-2.0
#PrettyTable>=0.7,<0.8
#kombu>=3.0.7
oslo.config
oslo.log
PrettyTable
kombu